agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents>=1.3.12",
    "livekit-plugins-openai>=1.3.12",
    "livekit-plugins-silero>=1.3.12",
    "python-dotenv>=1.2.1",
]
